backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic-settings==2.4.0
supabase==2.7.4
requests==2.32.3
httpx==0.27.2
python-dotenv==1.0.1
